+++ KaaTests/IPTransportInfoTest.m
@@ -21,7 +21,7 @@
     [KeyUtils generateKeyPair];
     NSData *publicKey = [KeyUtils getPublicKey];
     TransportProtocolId *TPid = [TransportProtocolIdHolder TCPTransportID];
-    int port = 80;
+    uint32_t port = 80;
     
     IPTransportInfo *info = [[IPTransportInfo alloc] initWithTransportInfo:[self createTestServerInfoWithServerType:SERVER_OPERATIONS transportProtocolId:TPid host:@"localhost" port:port andPublicKey:publicKey]];
     
@@ -31,7 +31,7 @@
     XCTAssertEqual(port, [info getPort]);
 }
 
-- (id<TransportConnectionInfo>) createTestServerInfoWithServerType:(ServerType)serverType transportProtocolId:(TransportProtocolId *)TPid host:(NSString *)host port:(NSUInteger)port andPublicKey:(NSData *)publicKey {
+- (id<TransportConnectionInfo>) createTestServerInfoWithServerType:(ServerType)serverType transportProtocolId:(TransportProtocolId *)TPid host:(NSString *)host port:(uint32_t)port andPublicKey:(NSData *)publicKey {
     ProtocolMetaData *md = [[ProtocolMetaData alloc] init];
     md = [self buildMetaDataWithTPid:TPid host:host port:port andPublicKey:publicKey];
     return  [[GenericTransportInfo alloc] initWithServerType:serverType andMeta:md];
@@ -39,10 +39,10 @@
 
 - (ProtocolMetaData *) buildMetaDataWithTPid:(TransportProtocolId *)TPid
                                         host:(NSString *)host
-                                        port:(NSUInteger)port
+                                        port:(uint32_t)port
                                 andPublicKey:(NSData *)publicKey {
-    NSUInteger publicKeyLength = [publicKey length];
-    NSUInteger hostLength = [host lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
+    uint32_t publicKeyLength = [publicKey length];
+    uint32_t hostLength = [host lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
     NSMutableData *data = [NSMutableData data];
     ProtocolVersionPair *pair = [[ProtocolVersionPair alloc]init];
     [pair setId:TPid.protocolId];
